refactor(cli): extract parser value enums
